# swagger/api/openapi.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - username
                - email
                - password
              example:
                username: "newuser"
                email: "new-user@email.com"
                password: "NewUser123"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      username:
                        type: string
                      email:
                        type: string
                      password:
                        type: string
                      updated_at:
                        type: string
                      created_at:
                        type: string
                      id:
                        type: integer
                  auth:
                    type: object
                    properties:
                      token:
                        type: string
                      type:
                        type: string
                example:
                  status: "Success"
                  message: "User created successfully"
                  user:
                    username: "newuser"
                    email: "new-user@email.com"
                    password: "$2y$12$tjwddwXp6E/87dkdqokmm8@$)d..."
                    updated_at: "2024-09-01T13:55:46.000000Z"
                    created_at: "2024-09-01T13:55:46.000000Z"
                    id: 2
                  auth:
                    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    type: "Bearer"
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: "Method Not Allowed"
        '422':
          description: Validation errors
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  errors:
                    type: array
                    items:
                      type: string
                example:
                  message: "The username has already been taken. (and 2 more errors)"
                  errors:
                    username:
                      - "The username has already been taken."
                    email:
                      - "The email field is required."
                    password:
                      - "The password field must be at least 6 characters"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT